+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-\-
+### Added
+- `Game.regions` property.
 
 ## [0.8.0] - 2022-05-11
 ### Added